Visualizing audio with resampled
