+++ deploy/client/src/components/AudioVisualizer.js
@@ -5,10 +5,10 @@
 */
 
 import React, { useEffect, useRef } from "react";
-import { useSourceAnalyser, useByteTimeDomainDataRef } from "../hooks";
+import { useSourceAnalyser } from "../hooks";
 import { useAudioStreamSource } from "../providers/AudioStreamSource";
 
-function sampleMaxAmplitudes(input, outputLength) {
+function resampleMaxAmplitudes(input, outputLength) {
   const output = [];
   const sampleWidth = Math.floor(input.length / outputLength);
   for (let i = 0; i < outputLength; ++i) {
@@ -27,24 +27,50 @@ function sampleMaxAmplitudes(input, outputLength) {
 const AudioVisualizer = ({
   amplitudeSpacing = 3,
   amplitudeWidth = 2,
-  displayWidthInSeconds = 3,
+  displaySeconds = 3,
   strokeStyle = "#fff",
   ...props
 }) => {
   const { source } = useAudioStreamSource();
   const analyser = useSourceAnalyser(source);
-  const byteTimeDomainDataRef = useByteTimeDomainDataRef(
-    analyser,
-    displayWidthInSeconds
-  );
+  const amplitudesRef = useRef([]);
   const canvasRef = useRef();
 
   useEffect(() => {
-    const draw = () => {
-      if (!byteTimeDomainDataRef.current) {
-        return;
-      }
+    if (analyser) {
+      const fftSize = analyser.fftSize;
+      const sampleRate = analyser.context.sampleRate;
+      const width = canvasRef.current.width;
+      const amplitudesLength = Math.floor(
+        width / (amplitudeSpacing + amplitudeWidth)
+      );
+      const resampledLength = Math.floor(
+        (fftSize * amplitudesLength) / (sampleRate * displaySeconds)
+      );
 
+      let amplitudes = Array(amplitudesLength).fill(-Infinity);
+      const timeDomainData = new Uint8Array(fftSize);
+      const intervalId = setInterval(() => {
+        amplitudes = amplitudes.slice(resampledLength);
+        analyser.getByteTimeDomainData(timeDomainData);
+        const resampled = resampleMaxAmplitudes(
+          timeDomainData,
+          resampledLength
+        );
+        amplitudes.push(...resampled);
+        amplitudesRef.current = amplitudes;
+      }, (fftSize / sampleRate) * 1e3);
+
+      return function () {
+        clearInterval(intervalId);
+      };
+    }
+  }, [analyser, amplitudeSpacing, amplitudeWidth, canvasRef, displaySeconds]);
+
+  useEffect(() => {
+    let animationFrame;
+
+    const draw = () => {
       const canvas = canvasRef.current;
       const ctx = canvas.getContext("2d");
       ctx.lineWidth = amplitudeWidth;
@@ -53,24 +79,20 @@ const AudioVisualizer = ({
       const { height, width } = canvas;
       const midHeight = height / 2;
 
-      const amplitudes = sampleMaxAmplitudes(
-        byteTimeDomainDataRef.current,
-        width / (amplitudeSpacing + amplitudeWidth)
-      );
-
       let x = 0;
       ctx.clearRect(0, 0, width, height);
       ctx.beginPath();
-      for (const amplitude of amplitudes) {
-        const y = Math.max(2, (amplitude / 255.0) * height);
-        ctx.moveTo(x, midHeight - y / 2);
-        ctx.lineTo(x, midHeight + y / 2);
+      for (const amplitude of amplitudesRef.current) {
+        if (isFinite(amplitude)) {
+          const y = Math.max(2, (amplitude / 255.0) * height);
+          ctx.moveTo(x, midHeight - y / 2);
+          ctx.lineTo(x, midHeight + y / 2);
+        }
         x += amplitudeSpacing + amplitudeWidth;
       }
       ctx.stroke();
     };
 
-    let animationFrame;
     const render = () => {
       draw();
       animationFrame = requestAnimationFrame(render);
@@ -80,13 +102,7 @@ const AudioVisualizer = ({
     return () => {
       cancelAnimationFrame(animationFrame);
     };
-  }, [
-    amplitudeSpacing,
-    amplitudeWidth,
-    byteTimeDomainDataRef,
-    canvasRef,
-    strokeStyle,
-  ]);
+  }, [amplitudeSpacing, amplitudeWidth, canvasRef, strokeStyle]);
 
   return <canvas ref={canvasRef} {...props} />;
 };
