+++ deploy/client/src/hooks/index.js
@@ -1,3 +1,2 @@
 export { useAudioContext } from "./useAudioContext";
-export { useByteTimeDomainDataRef } from "./useByteTimeDomainDataRef";
 export { useSourceAnalyser } from "./useSourceAnalyser";
